agregar prueba loadcalendar sin mock
